linebot_with_django4_0/requirements.txt
Django>=4.*
line-bot-sdk>=2.3.0
gunicorn==20.*
gevent==21.*